+++ Team16/Actor/Bulletes/WallReflectionBullet.h
@@ -9,7 +9,7 @@
 class WallReflectionBullet : public BaseObject
 {
 public:
-	WallReflectionBullet(Vector2 pos, CharactorManager* c, Type t, float angle);
+	WallReflectionBullet(Vector2 pos, CharactorManager* c, Type t, int angle);
 	~WallReflectionBullet();
 	//弾のタイプ分け
 	void setBulletType();
@@ -33,10 +33,10 @@ private:
 	CharactorManager* charaManager;
 	Type b_SetType;
 	Vector2 playerPos;//プレイヤーのpos;
-	float angle;
+	int angle;
 	float PI = 3.141592653589793f;//円周率
-	float bulletAngle;			  //bulletの角度
-	Vector2 RotationZ(float ang);//z軸回転の成分だけ
+	int bulletAngle;			  //bulletの角度
+	Vector2 RotationZ(int ang);//z軸回転の成分だけ
 
 	WallType type;
 };
